Update test for status count
